nic_router: use jitterentropy for NAT range start
The NAT feature of nic_router always started the outgoing source port
with a fixed value of 49152. Consider following scenario:

client -> nic_router -> external_nat -> external_server
^                  ^
|                  |
+------Genode------+

If a client establishes a TCP connection to the external_server, and we
restart the Genode system, the client will connect to the server again.
However, from the server's perspective, the TCP session has not been
properly closed. Since the new connect will come from the same
src_ip/src_port combination, the server will forcefully reset the
connection. If that happens a lot, the server may even start to ignore
further connections from this ip/port combination for a while as a
mitigation. Since the nic_router always starts with the same port, even
the external_nat won't help, because its NAT-table will still contain
the entry with src_ip/src_port combination. And from the external_nat's
perspective, for each forwarded packet it will always refresh the
timeout for that entry.

This patch adds entropy to the dynamic port start of the NAT, so the
probability of running into this situation is minimized.

Issue #4086, gapfruit#1171