libc: add `arg-raw` argument type
There are special characters in XML [1]. If the node content or the attribute
value contains one of them, it is supposed to be escaped [1]. Now, while the
`xmllint` used by Genode is aware of those and follows XML standards, the Genode
XML parser treats it more relaxedly. As a result, if we need to pass a string
with an ampersand to a `libc` application, we can't. The string
`<arg value="&val"/>` is invalid from the perspective of the `xmllint`, by the
standard, we should instead write `<arg value="&arg;val"/>`. But Genode XML
parser will not decode sequence `&amp;` to an ampersand and the application will
receive `&arg;val` argument instead of `&val`.

This commit contains a workaround: it adds `arg-raw` node type. The content of
this node will be decoded and passed to the application. E.g.,if `config` has
line `<arg-raw>&amp;val<arg-raw/>` the application will receive decoded content
of the node: `&val`.

[1] https://en.wikipedia.org/wiki/List_of_XML_and_HTML_character_entity_referen
ces#Predefined_entities_in_XML

Issue gapfruit#1216