sandbox: enable CPU fault detection for children
Start nodes now support the sub node `<monitor_pd_faults/>`.
If this node is present, a fault handler is attached to the CPU-session
of the child.
If a CPU fault in a child is detected, and state reporting is
enabled, the attribute `cpu_fault` is set to true.

Issue gapfruit#1446