vfs/import: fix overwrite option
This patch fixes an issue in the overwrite feature where the source file
is smaller than the to be overwritten destination file. Only the data of
the source file were written which resulted in having garbage data in
the destination file.

This patch should not break existing scenarios, e.g. on file systems
where unlink is not implemented.

The other possible approach of using `ftruncate` resulted in
`FTRUNCATE_ERR_NO_PERM` on rump and on lxfs.

Issue gapfruit#1508