vfs/rump: add option to exit on fs error
The vfs server is resilient in a way that it catches all exceptions from
its plugins. When one plugin fails, the others still work. In the case
of rump being unable to mount a file system and act on it, with e.g.
fsck or mkfs, this is inconvenient. This patch adds the option to exit
when rump is unable to mount the partition.

Issue gapfruit#1570