depot_deploy/depot_query: depot raw access
This expands the content mechanism of `runtime` files, with the
possibility, to grant access to depot raw archive file system
content.
It is uses as follows:
```
  <content>
    <file_system label="<raw_archive_name>"/>
  </content>
```

For each `file_system` node a File_system route node after the
following pattern is created.
```
  <service name="File_system" label="<raw_archive_name>">
    <child name="depot_raw_chroot"
           label="<user> -> raw -> <name> -> <version>"/>
  </service>
```

The name of the target for the route is currently set fix to
`depot_raw_chroot`. This component needs to exist as a sibling
to the start node the runtime is started in.

Issue gapfruit#1386